transfer of loader ru-en language pack from LatestMessagesFragment to SignInFragment
